requirements.txt
alembic==1.4.2
astroid==2.4.2
autopep8==1.5.3
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
ply==3.11
psycopg2==2.8.5
pyasn1==0.4.8
pycodestyle==2.6.0
pycryptodomex==3.9.7
pylint==2.5.3
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-interface==1.6.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
